move device_mem_size function to compat file
